--- a/be/src/olap/uint24.h
+++ b/be/src/olap/uint24.h
@@ -140,9 +140,7 @@ public:
         return std::string(buf);
     }
 
-    const uint8_t* get_data() const {
-        return data;
-    }
+    const uint8_t* get_data() const { return data; }
 
 private:
     uint8_t data[3];