--- a/be/src/vec/columns/column_vector.h
+++ b/be/src/vec/columns/column_vector.h
@@ -170,7 +170,7 @@ public:
 
     void insert_date_column(const char* data_ptr, size_t num) {
         size_t input_value_size = sizeof(uint24_t);
-        
+
         for (int i = 0; i < num; i++) {
             uint64_t val = 0;
             memcpy((char*)(&val), data_ptr, input_value_size);