@wangbo 2022-04-12
#  Performance Test
## SQL1
```
SELECT sum(LO_EXTENDEDPRICE * LO_DISCOUNT) AS revenue
FROM lineorder_flat
WHERE LO_ORDERDATE >= 19930101 and LO_ORDERDATE <= 19931231 AND LO_DISCOUNT BETWEEN 1 AND 3 AND LO_QUANTITY < 25;

before:
- BlockLoadTime: 1s226ms
- PredColumnReadTime: 734.907ms
- VectorPredEvalTime: 296.752ms
- OutputColumnTime: 179.064ms

after:
- BlockLoadTime: 903.229ms
- PredColumnReadTime: 738.852ms
- VectorPredEvalTime: 83.223ms
- OutputColumnTime: 65.080ms
```

## SQL2:  pred only contains date
```
SELECT sum(LO_DISCOUNT) AS revenue
FROM lineorder_flat
WHERE LO_ORDERDATE >= 19930101 and LO_ORDERDATE <= 19931231

before:
- BlockLoadTime: 1s982ms
- PredColumnReadTime: 478.911ms
- VectorPredEvalTime: 307.892ms
- OutputColumnTime: 1s180ms

after:
- BlockLoadTime: 879.068ms
- PredColumnReadTime: 489.480ms
- VectorPredEvalTime: 100.503ms
- OutputColumnTime: 274.997ms

```

## SQL3: no predicate
```
select count(distinct LO_ORDERDATE) from lineorder_flat

before:
- BlockLoadTime: 8s395ms
- PredColumnReadTime: 8s002ms

after:
- BlockLoadTime: 6s641ms
- PredColumnReadTime: 6s252ms
```

We can see ```VectorPredEvalTime```  has significant performance improvement.